move `NewProjectType` enum so it can be used in the service
