+++ src/vs/workbench/browser/positronNewProjectWizard/interfaces/newProjectWizardEnums.ts
@@ -36,19 +36,6 @@ export enum PythonEnvironmentProvider {
 }
 
 /**
- * NewProjectType enum. Defines the types of projects that can be created.
- * TODO: localize. Since this is an enum, we can't use the localize function
- * because computed values must be numbers (not strings). So we'll probably need to
- * turn this into an object with keys and values, maybe also using something like
- * satisfies Readonly<Record<string, string>>.
- */
-export enum NewProjectType {
-	PythonProject = 'Python Project',
-	RProject = 'R Project',
-	JupyterNotebook = 'Jupyter Notebook'
-}
-
-/**
  * PythonRuntimeFilter enum.
  */
 export enum PythonRuntimeFilter {
