+++ src/vs/workbench/services/positronNewProject/common/positronNewProject.ts
@@ -43,6 +43,19 @@ export enum NewProjectStartupPhase {
 }
 
 /**
+ * NewProjectType enum. Defines the types of projects that can be created.
+ * TODO: localize. Since this is an enum, we can't use the localize function
+ * because computed values must be numbers (not strings). So we'll probably need to
+ * turn this into an object with keys and values, maybe also using something like
+ * satisfies Readonly<Record<string, string>>.
+ */
+export enum NewProjectType {
+	PythonProject = 'Python Project',
+	RProject = 'R Project',
+	JupyterNotebook = 'Jupyter Notebook'
+}
+
+/**
  * NewProjectTask enum. Defines the tasks that can be pending during new project initialization.
  */
 export enum NewProjectTask {
